requirements.txt
matplotlib==3.1.1
python-numpy==1.17.3
scipy==1.3.1
pandas==0.25.2